politips/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.6                # via pip-tools
decorator==4.0.10         # via ipython, traitlets
django-environ==0.4.0
django-extensions==1.6.7
django-filter==0.13.0
django-grappelli==2.8.1
Django==1.10              # via django-environ
djangorestframework-jwt==1.8.0
djangorestframework==3.4.3
first==2.0.1              # via pip-tools
ipdb==0.10.1
ipython-genutils==0.1.0   # via traitlets
ipython==5.0.0
Markdown==2.6.6
mysqlclient==1.3.7
neo4j-driver==1.0.2
pexpect==4.2.0            # via ipython
pickleshare==0.7.3        # via ipython
pip-tools==1.7.0
prompt-toolkit==1.0.5     # via ipython
psycopg2==2.6.2
ptyprocess==0.5.1         # via pexpect
pygments==2.1.3           # via ipython
PyJWT==1.4.2              # via djangorestframework-jwt
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via django-environ, django-extensions, pip-tools, prompt-toolkit
traitlets==4.2.2          # via ipython
uWSGI==2.0.13.1
wcwidth==0.1.7            # via prompt-toolkit
whitenoise==3.2

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via ipdb, ipython
